Resolve #1072: include code to allow Ender Arrows to use gateways.
This simply replicates the functionality of an Ender Pearl as was
suggested in the comments. I've honestly never actually used an ender
pearl from a distance before, and it took me a few tries to get the
arrow to hit the portal, but it seems to be functional properly in both
directions.